# locationAPI/nodejs/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "The ETSI MEC ISG MEC012 Location API described using OpenAPI. The\
    \ API is based on the Open Mobile Alliance's specification RESTful Network API\
    \ for Zonal Presence"
  version: "1.0.0"
  title: "Location API"
  license:
    name: "BSD-3-Clause"
    url: "https://forge.etsi.org/legal-matters"
host: "127.0.0.1:8081"
basePath: "/exampleAPI/location/v1"
...
schemes:
- "http"
- "https"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /zones/{zoneId}:
    get:
      tags:
      - "zones"
      description: "Used to get the status of a zone."
      operationId: "zonesGetById"
      produces:
      - "application/json"
      parameters:
      - name: "zoneId"
        in: "path"
        description: "Zone ID"
        required: true
        type: "string"
      responses:
        "200":
          description: "Successful response to a query regarding the status of a zone"
          examples:
            application/json:
              zoneInfo:
                zoneId: "zone01"
                numberOfAccessPoints: "3"
                numberOfUnserviceableAccessPoints: "1"
                numberOfUsers: "10"
                resourceURL: "http://example.com/exampleAPI/location/v1/zones/zone01"
          schema:
            properties:
              zoneInfo:
                $ref: "#/definitions/ZoneInfo"
      x-swagger-router-controller: "Zones"
  /zones/{zoneId}/accessPoints:
    get:
      tags:
      - "zones"
      description: "Access point status can be retrieved for sets of access points\
        \ matching attribute in the request."
      operationId: "zonesByIdGetAps"
      produces:
      - "application/json"
      parameters:
      - name: "zoneId"
        in: "path"
        description: "Zone ID"
        required: true
        type: "string"
      - name: "interestRealm"
        in: "query"
        description: "Interest realm of access point (e.g. geographical area, a type\
          \ of industry etc.)."
        required: false
        type: "string"
      responses:
        "200":
          description: "Successful response to a query a named set of access point\
            \ status request"
          examples:
            application/json:
              accessPointList:
                zoneId: "zone01"
                accessPoint:
                - accessPointId: "001010000000000000000000000000001"
                  locationInfo:
                    latitude: "90.123"
                    longitude: "80.123"
                    altitude: "10.0"
                    accuracy: "0"
                  connectionType: "Macro"
                  operationStatus: "Serviceable"
                  numberOfUsers: "5"
                  interestRealm: "LA"
                  resourceURL: "http://example.com/exampleAPI/location/v1/zones/zone01/accessPoints/ap001"
                - accessPointId: "001010000000000000000000000000010"
                  locationInfo:
                    latitude: "91.123"
                    longitude: "81.123"
                    altitude: "12.0"
                    accuracy: "1"
                  connectionType: "Macro"
                  operationStatus: "Unserviceable"
                  numberOfUsers: "0"
                  interestRealm: "DC"
                  resourceURL: "http://example.com/exampleAPI/location/v1/zones/zone01/accessPoints/ap002"
                - accessPointId: "001010000000000000000000000000011"
                  locationInfo:
                    latitude: "93.123"
                    longitude: "83.123"
                    altitude: "16.0"
                    accuracy: "3"
                  connectionType: "Macro"
                  operationStatus: "Serviceable"
                  numberOfUsers: "5"
                  interestRealm: "NJ"
                  resourceURL: "http://example.com/exampleAPI/location/v1/zones/zone01/accessPoints/ap003"
                resourceURL: "http://example.com/exampleAPI/location/v1/zones/zone01/accessPoints"
          schema:
            properties:
              accessPointList:
                $ref: "#/definitions/AccessPointList"
      x-swagger-router-controller: "Zones"
  /zones/{zoneId}/accessPoints/{accessPointId}:
    get:
      tags:
      - "zones"
      description: "Access point status can be retrieved for sets of access points\
        \ matching attribute in the request."
      operationId: "zonesByIdGetApsById"
      produces:
      - "application/json"
      parameters:
      - name: "zoneId"
        in: "path"
        description: "Zone ID"
        required: true
        type: "string"
      - name: "accessPointId"
        in: "path"
        description: "Access Point ID"
        required: true
        type: "string"
      responses:
        "200":
          description: "Successful response to a query a named set of access point\
            \ status request"
          examples:
            application/json:
              accessPointInfo:
                accessPointId: "001010000000000000000000000000001"
                locationInfo:
                  latitude: "90.123"
                  longitude: "80.123"
                  altitude: "10.0"
                  accuracy: "0"
                connectionType: "Macro"
                operationStatus: "Serviceable"
                numberOfUsers: "5"
                interestRealm: "LA"
                resourceURL: "http://example.com/exampleAPI/location/v1/zones/zone001/accessPoints/ap001"
          schema:
            properties:
              accessPointInfo:
                $ref: "#/definitions/AccessPointInfo"
      x-swagger-router-controller: "Zones"
  /users:
    get:
      tags:
      - "users"
      description: "Users currently using a zone may be retrieved for sets of access\
        \ points matching attribute in the request"
      operationId: "usersGet"
      produces:
      - "application/json"
      parameters:
      - name: "zoneId"
        in: "query"
        description: "Zone ID"
        required: true
        type: "string"
      - name: "accessPointId"
        in: "query"
        description: "Identifier of access point, reference \"definitions\" for string\
          \ format"
        required: false
        type: "string"
      responses:
        "200":
          description: "Successful response to a query users within a zone request"
          examples:
            application/json:
              userList:
                user:
                - address: "acr:192.0.2.1"
                  accessPointId: "001010000000000000000000000000001"
                  zoneId: "zone01"
                  resourceURL: "http://example.com/exampleAPI/location/v1/users/acr%3A192.0.2.1"
                - address: "acr:192.0.2.2"
                  accessPointId: "001010000000000000000000000000001"
                  zoneId: "zone01"
                  resourceURL: "http://example.com/exampleAPI/location/v1/users/acr%3A192.0.2.2"
                - address: "acr:192.0.2.3"
                  accessPointId: "001010000000000000000000000000010"
                  zoneId: "zone01"
                  resourceURL: "http://example.com/exampleAPI/location/v1/users/acr%3A192.0.2.3"
                - address: "acr:192.0.2.4"
                  accessPointId: "001010000000000000000000000000001"
                  zoneId: "zone02"
                  resourceURL: "http://example.com/exampleAPI/location/v1/users/acr%3A192.0.2.4"
                - address: "acr:192.0.2.5"
                  accessPointId: "001010000000000000000000000000010"
                  zoneId: "zone02"
                  resourceURL: "http://example.com/exampleAPI/location/v1/users/acr%3A192.0.2.5"
                resourceURL: "http://example.com/exampleAPI/location/v1/users"
          schema:
            properties:
              userList:
                $ref: "#/definitions/UserList"
      x-swagger-router-controller: "Users"
  /users/{userId}:
    get:
      tags:
      - "users"
      description: "Users currently using a zone may be retrieved for sets of access\
        \ points matching attribute in the request"
      operationId: "usersGetById"
      produces:
      - "application/json"
      parameters:
      - name: "userId"
        in: "path"
        description: "User ID"
        required: true
        type: "string"
      responses:
        "200":
          description: "Successful response to a query users within a zone request"
          examples:
            application/json:
              userInfo:
                address: "acr:192.0.2.1"
                accessPointId: "001010000000000000000000000000001"
                zoneId: "zone01"
                resourceURL: "http://example.com/exampleAPI/location/v1/users/acr%3A192.0.2.1"
                locationInfo:
                  latitude: "90.123"
                  longitude: "80.123"
                  altitude: "10.0"
                  accuracy: "0"
                contextLocationInfo: "GroundFloor"
          schema:
            properties:
              userInfo:
                $ref: "#/definitions/UserInfo"
      x-swagger-router-controller: "Users"
...
definitions:
  AccessPointId:
    type: "string"
    description: "Identifier of access point, <E-CGI><Cell Portion ID> (reference\
      \ ETSI TS 129 171). Where the E-CGI is made up of the PLMN and Cell Identity\
      \ (28 bit string). Then the PLMN is made up of the 3 digit MCC & 2 or 3 digit\
      \ MNC. The Cell Portion is an optional element"
    example: "001010000000000000000000000000001"
  AccessPointInfo:
    type: "object"
    required:
    - "accessPointId"
    - "connectionType"
    - "numberOfUsers"
    - "operationStatus"
    - "resourceURL"
    properties:
      accessPointId:
        $ref: "#/definitions/AccessPointId"
      locationInfo:
        $ref: "#/definitions/LocationInfo"
      connectionType:
        $ref: "#/definitions/ConnectionType"
      operationStatus:
        $ref: "#/definitions/OperationStatus"
      numberOfUsers:
        $ref: "#/definitions/NumberOfUsers"
      timezone:
        $ref: "#/definitions/Timezone"
      interestRealm:
        $ref: "#/definitions/InterestRealm"
      resourceURL:
        $ref: "#/definitions/ResourceURL"
    description: "A type containing access point information."
  AccessPointList:
    type: "object"
    required:
    - "resourceURL"
    - "zoneId"
    properties:
      zoneId:
        $ref: "#/definitions/ZoneId"
      accessPoint:
        type: "array"
        description: "Collection of the access point information list."
        items:
          $ref: "#/definitions/AccessPointInfo"
      resourceURL:
        $ref: "#/definitions/ResourceURL"
    description: "A type containing list of access points."
  Address:
    type: "string"
    format: "uri"
    description: "Address of user (e.g. \"sip\" URI, \"tel\" URI, \"acr\" URI)."
    example: "acr:192.0.2.1"
  AncillaryInfo:
    type: "string"
    description: "Reserved for future use."
...
  ConnectionType:
    type: "string"
    description: "The connection type for the access point"
    example: "Macro"
    enum:
    - "Femto"
    - "LTE-femto"
    - "Smallcell"
    - "LTE-smallcell"
    - "Wifi"
    - "Pico"
    - "Micro"
    - "Macro"
    - "Wimax"
    - "Unknown"
  ContextLocationInfo:
    type: "string"
    description: "Contextual information of a user location (e.g., aisle, floor, room\
      \ number, etc.)"
    example: "GroundFloor"
...
  InterestRealm:
    type: "string"
    description: "Interest realm of access point (e.g. geographical area, a type of\
      \ industry etc.)."
    example: "LA"
...
  LocationInfo:
    type: "object"
    required:
    - "accuracy"
    - "latitude"
    - "longitude"
    properties:
      latitude:
        type: "number"
        format: "float"
        example: 80.123
      longitude:
        type: "number"
        format: "float"
        example: 70.123
      altitude:
        type: "number"
        format: "float"
        example: 10.0
      accuracy:
        type: "integer"
        format: "int32"
        example: 10
    description: "A type containing location information with latitude, longitude\
      \ and altitude, in addition the accuracy of the information are provided."
...
  NumberOfAccessPoints:
    type: "integer"
    format: "uint32"
    description: "The number of access points within the zone"
    example: "10"
  NumberOfUnserviceableAccessPoints:
    type: "integer"
    format: "uint32"
    description: "Number of inoperable access points within the zone."
    example: "9"
  NumberOfUsers:
    type: "integer"
    format: "uint32"
    description: "The number of users currently on the access point."
    example: "7"
...
  OperationStatus:
    type: "string"
    description: "The operation status of the access point"
    example: "Serviceable"
    enum:
    - "Serviceable"
    - "Unserviceable"
    - "Unknown"
...
  ResourceURL:
    type: "string"
    format: "uri"
    description: "Self referring URL."
    example: "http://example.com/exampleAPI/location/v1/subscriptions/userTracking/subscription123"
...
  Timezone:
    type: "string"
    format: "date-time"
    description: "Time zone of access point"
    example: "2017-01-01T02:51:43Z"
...
  UserInfo:
    type: "object"
    required:
    - "accessPointId"
    - "address"
    - "resourceURL"
    - "zoneId"
    properties:
      address:
        $ref: "#/definitions/Address"
      accessPointId:
        $ref: "#/definitions/AccessPointId"
      zoneId:
        $ref: "#/definitions/ZoneId"
      resourceURL:
        $ref: "#/definitions/ResourceURL"
      locationInfo:
        $ref: "#/definitions/LocationInfo"
      contextLocationInfo:
        $ref: "#/definitions/ContextLocationInfo"
      ancillaryInfo:
        $ref: "#/definitions/AncillaryInfo"
    description: "A type containing user information."
  UserList:
    type: "object"
    required:
    - "resourceURL"
    properties:
      user:
        type: "array"
        description: "Collection of the zone information list."
        items:
          $ref: "#/definitions/UserInfo"
      resourceURL:
        $ref: "#/definitions/ResourceURL"
    description: "A type containing list of users."
...
  ZoneId:
    type: "string"
    description: "Identifier of zone"
    example: "zone01"
  ZoneInfo:
    type: "object"
    required:
    - "numberOfAccessPoints"
    - "numberOfUnservicableAccessPoints"
    - "numberOfUsers"
    - "resourceURL"
    - "zoneId"
    properties:
      zoneId:
        $ref: "#/definitions/ZoneId"
      numberOfAccessPoints:
        $ref: "#/definitions/NumberOfAccessPoints"
      numberOfUnservicableAccessPoints:
        $ref: "#/definitions/NumberOfUnserviceableAccessPoints"
      numberOfUsers:
        $ref: "#/definitions/NumberOfUsers"
      resourceURL:
        $ref: "#/definitions/ResourceURL"
    description: "A type containing zone information."
...